4893 - extract logic into hook
